[refactor] extract classes from config.kt into separate files
